refactor(work): stop re-exporting coding shell
